When viewing all students, fetch student donations before rendering the page
The students page displays student summaries. Each student summary
displays donation information for each student. As of this commit, there
does not appear to be a way to ask a collection to also fetch model
collections (ideally we could specify which ones as well).

This is a temporary solution while we trying to discover the idiomatic
way to ask all models of a collection to fetch their own collections.