Fix bug causing NaN to be displayed as a goal value
When editing or creating a student using the student form component, if
the value of goal was ever null or undefined, the string NaN would be
displayed in goal's form field.

To fix this, we only include goal's value in the form field's value when
goal's value is truthy.

We also added a $ pretty format to make it more obvious what type of
data goes into the goal field.