When signing out from the home page, reload the page
In all other cases we use router.navigate('/') to load the home page;
however, if the user is on the home page already, the router says "hey,
I'm already here, no need to go anywhere!" and doesn't re-render the
page. This results in the Sign Out link remaining Sign Out, instead of
becoming Sign In after re-render.

By using reload instead of navigate, the page is re-rendered.